services/whatsapp_agent/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
chromadb>=0.4.0
langchain>=0.1.0
transformers>=4.30.0
torch>=2.0.0
python-multipart>=0.0.6
python-dotenv>=0.19.0
requests>=2.28.0
beautifulsoup4>=4.9.3
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
whisper-openai>=20231117
pydantic>=2.0.0 